callcenter-analytics/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-docx>=0.8.11
reportlab>=4.0.0
Jinja2>=3.1.0
streamlit>=1.28.0
pyyaml>=6.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
plotly>=5.18.0
